#2: Moves db functionality to separate file.
